refactor: CartRepository에서 Product를 직접 받도록 수정
`CartRepository.addCartProduct` 메서드가 `CartProduct` 대신 `Product`를 인자로 받도록 변경되었습니다. 이로 인해 `Product`를 `CartProduct`로 변환하는 책임이 Repository 구현체로 이동되었습니다.

또한, `deleteCartProduct` 메서드의 ID 타입을 `Int`에서 `Long`으로 수정하여 데이터베이스 스키마와 일관성을 맞췄습니다.

+++ app/src/main/java/woowacourse/shopping/data/CartRepositoryImpl.kt
@@ -16,7 +16,7 @@ class CartRepositoryImpl(
 
     override suspend fun getAllCartProducts(): List<CartProduct> = dao.getAll().map { it.toModel() }
 
-    override suspend fun deleteCartProduct(id: Int) {
+    override suspend fun deleteCartProduct(id: Long) {
         dao.delete(id.toLong())
     }
 }
